the attached patch is a comparison of the post_process.cpp in the /trunk folder
the rewrite of the post_process.cpp is very welcome. however, as it
wasn't proceeded through Git, we have no track of changes from the
previous version (as well as it happened since 2004, 2005 and other
versions). the patch file is the least thing we can do to show
graphically the major changes for those users that want to understand
the changes and maybe do "cherry-picking" rather than global pull of
the changes.